✨ Refactor maze components: remove unused files, implement new Canvas and Maze classes, and enhance pathfinding logic.
