+++ src/component/maze-generation/dfs/cell.ts
@@ -1,190 +1,11 @@
-import { Stack } from './Stack';
-
-const enum State {
-  MakingGrid,
-  FindingPath,
-  Rest,
-}
-
-class Rat {
-  private readonly grid: Cell[][];
-  private currentCell: Cell;
-  private visitedCells: Set<Cell> = new Set();
-  private cellStack: Stack<Cell> = new Stack();
-  private path: Cell[] = [];
-
-  constructor(grid: Cell[][]) {
-    this.grid = grid;
-    this.currentCell = this.grid[0][0];
-    this.cellStack.push(this.currentCell);
-  }
-
-  showGrid(ctx: CanvasRenderingContext2D) {
-    const rows = this.grid.length;
-    const cols = this.grid[0].length;
-    for (let r = 0; r < rows; r++) {
-      for (let c = 0; c < cols; c++) {
-        this.grid[r][c].show(600, rows, cols, ctx);
-      }
-    }
-  }
-
-  step() {
-    this.path.push(this.currentCell);
-    const n = this.grid.length;
-
-    if (this.currentCell === this.grid[n - 1][n - 1]) {
-      return;
-    }
-
-    const nextCell = this.cellStack.pop();
-    if (!nextCell) return;
-
-    this.currentCell = nextCell;
-
-    this.visitedCells.add(this.currentCell);
-    const unvisitedNeighbors = this.currentCell.neighbors.filter(
-      (cell) => !this.visitedCells.has(cell),
-    );
-    this.cellStack.push(...unvisitedNeighbors);
-  }
-
-  draw(ctx: CanvasRenderingContext2D) {
-    this.showGrid(ctx);
-    this.currentCell.highlight(this.grid[0].length, ctx);
-    this.path.forEach((cell) => {
-      cell.highlight(this.grid[0].length, ctx, '#1289A7');
-    });
-  }
-}
-
-let current: Cell;
-export class Maze {
-  size: number;
-  rows: number;
-  columns: number;
-  private readonly grid: Cell[][] = [];
-  stack: Stack<Cell>;
-  current: Cell;
-  currentState: State = State.MakingGrid;
-  rat: Rat;
-
-  constructor(size: number, rows: number, columns: number) {
-    this.size = size;
-    this.rows = rows;
-    this.columns = columns;
-    this.stack = new Stack<Cell>();
-    this.initGrid();
-
-    this.rat = new Rat(this.grid);
-  }
-
-  //@setup() -> setup Grid on the canvas and draw each individual cell
-  initGrid() {
-    for (let r = 0; r < this.rows; r++) {
-      let row = [];
-      for (let c = 0; c < this.columns; c++) {
-        let cell = new Cell(r, c, this.grid, this.size);
-        row.push(cell);
-      }
-      this.grid.push(row);
-    }
-    current = this.grid[0][0];
-  }
-
-  renderPathFind(ctx: CanvasRenderingContext2D, maze: HTMLCanvasElement) {
-    this.rat.step();
-    this.rat.draw(ctx);
-  }
-
-  renderMakeGrid(ctx: CanvasRenderingContext2D, maze: HTMLCanvasElement) {
-    // Set the first cell as visited
-    current.visited = true;
-    // Loop through the 2d grid array and call the show method for each cell instance
-    for (let r = 0; r < this.rows; r++) {
-      for (let c = 0; c < this.columns; c++) {
-        let grid = this.grid;
-        grid[r][c].show(this.size, this.rows, this.columns, ctx);
-      }
-    }
-
-    // This function will assign the variable 'next' to random cell out of the current cells available neighboring cells
-    let next = current.checkNeighbours();
-    // If there is a non visited neighbor cell
-    if (next) {
-      next.visited = true;
-      // Add the current cell to the stack for backtracking
-      this.stack.push(current);
-      // this function will highlight the current cell on the grid. The parameter columns is passed
-      // in order to set the size of the cell
-      current.highlight(this.columns, ctx);
-      // This function compares the current cell to the next cell and removes the relevant walls for each cell
-      current.removeWalls(current, next);
-      // Set the next cell to the current cell
-      current = next;
-
-      // Else if there are no available neighbours start backtracking using the stack
-    } else if (this.stack.length > 0) {
-      let cell = this.stack.pop();
-      current = cell;
-      current.highlight(this.columns, ctx);
-    }
-    // If no more items in the stack then all cells have been visited and the function can be exited
-    if (this.stack.length === 0) {
-      this.currentState = State.FindingPath;
-      return;
-    }
-  }
-
-  draw_utils(ctx: CanvasRenderingContext2D, maze: HTMLCanvasElement) {
-    const step = () => {
-      maze.width = this.size;
-      maze.height = this.size;
-      maze.style.background = 'black';
-
-      if (this.currentState === State.MakingGrid) {
-        this.renderMakeGrid(ctx, maze);
-      } else {
-        this.renderPathFind(ctx, maze);
-      }
-
-      window.requestAnimationFrame(step);
-    };
-
-    // Recursively call the draw function. This will be called up until the stack is empty
-    // window.requestAnimationFrame(() => {
-    //   this.draw_utils(ctx, maze);
-    // });
-    step();
-  }
-
-  // Draw the canvas by setting the size and placing the cells in the grid array on the canvas.
-
-  draw(ctx: CanvasRenderingContext2D, maze: HTMLCanvasElement): Promise<any> {
-    return new Promise<any>((_resolve, _reject): void => {
-      _resolve(this.draw_utils(ctx, maze));
-    });
-  }
-
-  traverse() {
-    // while (true) {
-    //     current = this.stack.top();
-    // }
-    console.log('Callback! Successful...');
-  }
-}
+import { Walls } from '@/types/types';
 
 export class Cell {
   rowNum: number;
   parentGrid: any[];
   parentSize: number;
   visited: boolean;
-  walls: {
-    topWall: boolean;
-    rightWall: boolean;
-    bottomWall: boolean;
-    leftWall: boolean;
-  };
+  walls: Walls;
   colNum: number;
 
   constructor(rowNum: number, columnNum: number, parentGrid: Array<any>, parentSize: number) {
@@ -199,8 +20,7 @@ export class Cell {
       bottomWall: true,
       leftWall: true,
     };
-    // parentGrid is passed in to enable the checkneighbours method.
-    // parentSize is passed in to set the size of each cell on the grid
+
     this.parentGrid = parentGrid;
     this.parentSize = parentSize;
   }
