+++ src/config/routes.tsx
@@ -1,16 +1,16 @@
-import { lazy } from "react";
-import type { RouteObject } from "react-router-dom";
+import { lazy } from 'react';
+import type { RouteObject } from 'react-router-dom';
 
-const Home = lazy(() => import("@/pages/Home"));
-const DFS = lazy(() => import("@/pages/Dfs"));
+const Home = lazy(() => import('@/pages/Home'));
+const DFS = lazy(() => import('@/pages/Dfs'));
 
 export const appRoutes: RouteObject[] = [
   {
-    path: "/",
+    path: '/',
     element: <Home />,
   },
   {
-    path: "/dfs",
+    path: '/dfs',
     element: <DFS />,
   },
 ];
