+++ src/data-structure/stack.ts
