+++ src/types/types.ts
@@ -1,3 +1,9 @@
+export const enum State {
+  MakingGrid,
+  FindingPath,
+  Rest,
+}
+
 export interface Walls {
   topWall: boolean;
   rightWall: boolean;
